requirements.txt
matplotlib.pyplot
numpy
Pillow
torch>=1.2
tabulate
tqdm
torchvision